% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{banner}
\alias{banner}
\title{banner}
\usage{
banner(txt, Char = "-")
}
\arguments{
\item{txt}{text to print in banner}

\item{Char}{character seperator (eg. "-", "=", "*")}
}
\description{
print banner as header with separator
}
