% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlots.R
\name{boxPlots}
\alias{boxPlots}
\title{boxPlots}
\usage{
boxPlots(data = data, group = group, drop.var = NULL, path = NULL)
}
\arguments{
\item{data}{normalized data}

\item{group}{grouping variables}

\item{drop.var}{drop un-necessary variables}

\item{path}{saving path}
}
\description{
boxPlots
}
