% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicalMetadata.R, R/chemicalMetadata_MH_mmu.R
\docType{data}
\name{chemicalMetadata}
\alias{chemicalMetadata}
\title{chemical metadata for annotation of metabolon}
\format{
\describe{
        \item{CHEM_ID}{Chemical id}
        \item{MET_CHEM_NO}{Chemical id}
        \item{LIB_ID}{library id}
        \item{COMP_ID}{CCompound id}
        \item{CHRO_LIB_ENTRY_ID}{CHRO_LIB_ENTRY_ID}
        \item{SUPER_PATHWAY}{ONTOLOGY CLASS}
        \item{SUB_PATHWAY}{ONTOLOGY SUBCLASS}
        \item{PATHWAY_SORTORDER}{pathway sorting}
        \item{TYPE}{Named / unnamed}
        \item{INCHIKEY}{INCHIKEY}
        \item{SMILES}{SMILES}
        \item{CHEMICAL_NAME}{Chemical name}
        \item{PLOT_NAME}{Chemical name}
        \item{CAS}{CAS id}
        \item{CHEMSPIDER}{Chemspider id}
        \item{HMDB}{HMDB id}
        \item{KEGG}{KEGG id}
        \item{PUBCHEM}{PUBCHEM id}
        \item{PLATFORM}{used platform}
         }

\describe{
        \item{MET_CHEM_NO}{Metabolite chem Name}
        \item{ONTOLOGY1_NAME}{Major ontology class}
        \item{ONTOLOGY2_NAME}{Ontology subclass}
        \item{METABOLITE_NAME}{Metabolite name}
        \item{PUBCHEM_ID}{Pubchem id}
        \item{Compound_ID}{compound id}
        \item{CAS_UID}{cas id}
        \item{HMDB_ID}{HMDB id}
        \item{KEGG_ID}{Kegg id}
         }
}
\source{
{metabolon}

{mh}
}
\usage{
chemicalMetadata

chemicalMetadata
}
\description{
contains a dataframe with metadata

contains a dataframe with metadata
}
\examples{
data(chemicalMetadata)
data(chemicalMetadata_MH_mmu)
}
\keyword{datasets}
