% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicalMetadata_MH.R
\docType{data}
\name{chemicalMetadata_MH}
\alias{chemicalMetadata_MH}
\title{chemical metadata for annotation of metanomics health}
\format{
\describe{
        \item{MET_CHEM_NO}{Chemical id}
        \item{UNIQUE_ANALYTE_ID}{unique ID}
        \item{ASSAY}{Assay Name}
        \item{METABOLITE_NAME}{metabolite name}
        \item{ONTOLOGY1_NAME}{ONTOLOGY CLASS}
        \item{ONTOLOGY2_NAME}{ONTOLOGY SUBCLASS}
        \item{CAS_UID}{CAS ID}
        \item{DICT_STRUCTURE_ELUCIDATION_ID}{structurre}
        \item{KEGG}{KEGG ID}
         }
}
\source{
{metanomics health}
}
\usage{
chemicalMetadata_MH
}
\description{
contains a dataframe with metadata
}
\examples{
data(chemicalMetadata_MH)
}
\keyword{datasets}
