% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{correlationPlot}
\usage{
correlationPlot(
  dataList,
  data.type = c("MH", "Metabolon", "Others"),
  species = c("hsa", "mmu"),
  group = NULL,
  var.imp = NULL
)
}
\arguments{
\item{dataList}{normalized data}

\item{data.type}{select platform used, c("MH", "Metabolon", "Others")}

\item{species}{species to use "hsa" or "mmu"}

\item{group}{grouping variables}
}
\value{
correlation matrix as a list and pdf output in defined path folder.
}
\description{
function to plot intragroup correlation plot.
}
