% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffAbundanceScore.R
\name{diffAbundanceScore}
\alias{diffAbundanceScore}
\title{diffAbundanceScore}
\usage{
diffAbundanceScore(
  species = c("hsa", "mmu"),
  ref.path = NULL,
  results,
  p.value.cutoff = 0.05,
  fold.changes.cutoff = 1.5,
  common.mets = 15,
  save = c("pdf", "svg", "png"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300
)
}
\arguments{
\item{species}{species to use "hsa" or "mmu"}

\item{ref.path}{saving path}

\item{results}{fold changes results}

\item{p.value.cutoff}{cutoff value of p.value}

\item{fold.changes.cutoff}{higher cutoff value of fold changes}

\item{common.mets}{minimum number of common metaolites}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}
}
\value{
output in defined path
}
\description{
The differential abundance (DA) score captures the tendency for a pathway t
o have increased levels of metabolites, relative to a control group.
The score is calculating by first applying a non-parametric differential
abundance test (in this study, Benjamini-Hochberg corrected Mann-Whitney
U-tests) to all metabolites in a pathway. Then, after determining which
metabolites are significantly increased/decreased in abundance, the
differential abundance score is defined as:
$$
DAS = {n(Metabolites_{up}) - n(Metabolites_{down})}{n(Metabolites_{up}) + n(Metabolites_{down})}
$$
Thus, the DA score varies from -1 to 1. A score of -1 indicates that all
metabolites in a pathway decreased in abundance, while a score of 1
indicates that all metabolites increased.
}
