% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionPlot.R
\name{distributionPlot}
\alias{distributionPlot}
\title{distributionPlot}
\usage{
distributionPlot(
  data = data,
  path = NULL,
  cutoff = 0.01,
  lipid.class = TRUE,
  save = "pdf",
  fig.width = 12,
  fig.height = 9,
  dpi = 300,
  ...
)
}
\arguments{
\item{data}{fold changes data}

\item{path}{saving path}

\item{cutoff}{cutoff of p value}

\item{lipid.class}{TRUE/FALSE to plot lipid classes}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}

\item{...}{ggplot extensions}
}
\description{
distributionPlot
}
