% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichedNetwork.R
\name{enrichedNetwork}
\alias{enrichedNetwork}
\title{enrichedNetwork}
\usage{
enrichedNetwork(
  species = c("hsa", "mmu"),
  ref.path = NULL,
  results,
  p.value.cutoff = 0.05,
  fold.changes.cutoff = 1.5,
  network.method = "diffusion",
  legend.pathway = 3,
  save = c("pdf", "svg", "png", "none"),
  data.type = c("MH", "Metabolon", "Others"),
  fig.width = 12,
  fig.height = 9,
  Other_metadata = NULL
)
}
\arguments{
\item{species}{species to use "hsa" or "mmu"}

\item{ref.path}{saving path}

\item{results}{fold changes results}

\item{p.value.cutoff}{cutoff value of p.value}

\item{fold.changes.cutoff}{higher cutoff value of fold changes}

\item{network.method}{"diffusion" as method of choice, for other method refer to FELLA. CUrrently only caliberated for diffusion.}

\item{legend.pathway}{number of pathways to map on plot}

\item{save}{either "pdf", "svg", "png" or "none"}

\item{data.type}{select platform used, c("MH", "Metabolon", "Others")}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{Other_metadata}{dataframe with metadata....it must have  columns: Metabolite, Metabolite_Name, Ontology_Class, Ontology_Subclass.}
}
\value{
results of enrichment network and network plots as save object in defined path.
The object contains the following:\itemize{
    \item results enrichment results
    \item plot.pathway.impact list of all the plots of pathway.impact
  }
}
\description{
plot enriched annotated netwroks for the signficantly altered metabolic pathways.
}
