% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentScore.KEGG.R
\name{enrichmentScore.KEGG}
\alias{enrichmentScore.KEGG}
\title{enrichmentScore.KEGG}
\usage{
enrichmentScore.KEGG(
  species = c("hsa", "mmu"),
  ref.path = NULL,
  results,
  p.value.cutoff = 0.05,
  fold.changes.cutoff = 1.5,
  save = c("pdf", "svg", "png"),
  data.type = c("MH", "Metabolon", "Others"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300,
  Other_metadata = NULL
)
}
\arguments{
\item{species}{species to use "hsa" or "mmu"}

\item{ref.path}{saving path}

\item{results}{fold changes results}

\item{p.value.cutoff}{cutoff value of p.value}

\item{fold.changes.cutoff}{higher cutoff value of fold changes}

\item{save}{either "pdf", "svg" or "png"}

\item{data.type}{select platform used, c("MH", "Metabolon", "Others")}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}

\item{Other_metadata}{dataframe with metadata....it must have  columns: Metabolite, Metabolite_Name, Ontology_Class, Ontology_Subclass.}
}
\value{
data.frame onject with enrichement.results and plots as save object in defined path.
}
\description{
enrichmentScore calualtion based on enrichement pathway analysis.
}
