% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.R
\name{exportResults}
\alias{exportResults}
\title{exportResults}
\usage{
exportResults(data, path = NULL)
}
\arguments{
\item{data}{log fold changes data generated from normalizeDat function}

\item{path}{saving path of the data}
}
\value{
color coded result table in excel formant
}
\description{
export results with precolored formatting
}
