% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDiamReduction.R
\name{gg_circle}
\alias{gg_circle}
\title{gg_circle}
\usage{
gg_circle(rx, ry, xc, yc, color = "black", fill = NA, ...)
}
\arguments{
\item{rx}{x axis cordinates on r}

\item{ry}{y axis cordinates on r}

\item{xc}{x axis cordinate on c}

\item{yc}{y axis cordinate on c}

\item{color}{circle border color}

\item{fill}{circle fill}

\item{...}{ggplot extensions}
}
\description{
plot opls circle
}
