% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeTransformScale.R
\name{ImputeTransformScale}
\alias{ImputeTransformScale}
\title{ImputeTransformScale}
\usage{
ImputeTransformScale(
  Data,
  Impute = FALSE,
  Transform = FALSE,
  Scaling = FALSE,
  ScaleType = ScaleType
)
}
\arguments{
\item{Data}{data-frame with column names}

\item{Impute}{whether to perform median imputation, default is TRUE}

\item{Transform}{whether to perform log 10 transformation, default is TRUE}

\item{Scaling}{whether to perform scaling, default is TRUE}

\item{ScaleType}{Scaling method, either of "Centering", "Auto", "Range","Pareto", "Vast", "Level".}
}
\description{
This function median Imputation, transformation and scaling by different methods.
}
\keyword{Imputations,}
\keyword{Scaling}
\keyword{transformation,}
