% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installScriptLibs.R
\name{installScriptLibs}
\alias{installScriptLibs}
\title{installScriptLibs}
\usage{
installScriptLibs(packages)
}
\arguments{
\item{packages}{list of required packages}
}
\description{
function to install and load required listed packages
}
