% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leveneStat.R
\name{leveneStat}
\alias{leveneStat}
\title{leveneStat}
\usage{
leveneStat(
  group = group,
  data = data,
  location = "median",
  trim.alpha = 0.25,
  bootstrap = TRUE,
  num.bootstrap = 1000,
  kruskal.test = TRUE
)
}
\arguments{
\item{group}{grouping variable}

\item{data}{dataframe of normalized data}

\item{location}{mean or median, median is default}

\item{trim.alpha}{specify the percentage of trimmed mean}

\item{bootstrap}{whether to perform bootstrapping TRUE/FALSE}

\item{num.bootstrap}{number of bootstraping steps}

\item{kruskal.test}{whether to perform kruskal test TRUE/FALSE}
}
\description{
Computes Levene's stats for homogeneity of variance across groups
}
