% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidChainLengthCorrelation.R
\name{lipidChainLengthCorrelation}
\alias{lipidChainLengthCorrelation}
\title{lipidChainLengthCorrelation}
\usage{
lipidChainLengthCorrelation(
  results,
  path = NULL,
  save = c("pdf", "svg", "png"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300
)
}
\arguments{
\item{results}{fold changes data}

\item{path}{saving path}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}
}
\value{
plot as save object in defined path.
}
\description{
compute correlation of chain length with abundance.
}
