% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidChainLengthDistribution.R
\name{lipidChainLengthDistribution}
\alias{lipidChainLengthDistribution}
\title{lipidChainLengthDistribution}
\usage{
lipidChainLengthDistribution(
  results,
  species = c("hsa", "mmu"),
  p.value.cutoff = 0.05,
  fold.changes.cutoff = 1.5,
  path = NULL,
  save = c("pdf", "svg", "png"),
  data.type = c("MH", "Metabolon", "Others"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300,
  Other_metadata = NULL
)
}
\arguments{
\item{results}{fold changes data}

\item{species}{species to use "hsa" or "mmu"}

\item{p.value.cutoff}{cutoff of p-values to be used}

\item{fold.changes.cutoff}{higher fold changes cutoff to be used}

\item{path}{saving path}

\item{save}{either "pdf", "svg" or "png"}

\item{data.type}{select platform used, c("MH", "Metabolon", "Others")}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}

\item{Other_metadata}{dataframe with metadata....it must have  columns: Metabolite, Metabolite_Name, Ontology_Class, Ontology_Subclass}
}
\value{
plot in save object in defined path.
}
\description{
compute distribution of chain length per class
}
