% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeDat.binary.R
\name{normalizeDat.binary}
\alias{normalizeDat.binary}
\title{normalizeDat.binary}
\usage{
normalizeDat.binary(
  dataList,
  confounders = NULL,
  stratifier,
  fix.effect = NULL,
  random.effect = NULL
)
}
\arguments{
\item{dataList}{raw metabolome data list from imputeTransformScale function.It need to have imputed.matrix and metadata.}

\item{confounders}{list of confounders, NULL for mixed effect model}

\item{stratifier}{classifier variable of interest, Should not be in confounders. Should be one length 1.}

\item{fix.effect}{equation of fixed effect, NULL for fixed effect model}

\item{random.effect}{equation of random effects, NULL for fixed effect model}
}
\value{
Analyses results in list object.
  The object contains the following:\itemize{
    \item fitted.value fitted values for each variable
    \item summaryFC anova results
  }
}
\description{
normalize data using fixed effect as well as mixed effects models for comparison of binary groups ie group vs rest.
}
