% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oplsPlot.R
\name{oplsPlot}
\alias{oplsPlot}
\title{oplsPlot}
\usage{
oplsPlot(data = data, group = group, drop.grouping.var = NULL, ...)
}
\arguments{
\item{data}{metabolome data}

\item{group}{stratifying variable}

\item{drop.grouping.var}{drop grouping varibales not necessary}

\item{...}{ggplot extensions}
}
\description{
Partial Least-Squares Discriminant Analysis (PLS-DA) is a multivariate dimensionality-reduction tool.
}
