% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPlot.R
\name{pcaPlot}
\alias{pcaPlot}
\title{pcaPlot}
\usage{
pcaPlot(data = data, group = group, drop.grouping.var = NULL, ...)
}
\arguments{
\item{data}{metabolome dta}

\item{group}{stratifying variable}

\item{drop.grouping.var}{drop grouping varibales not necessary for PCA}

\item{...}{ggplot extensions}
}
\description{
Principle component analysis
}
