% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piePlot.R
\name{piePlot}
\alias{piePlot}
\title{piePlot}
\usage{
piePlot(
  data = data,
  path = NULL,
  cutoff = 0.05,
  save = "pdf",
  fig.width = 12,
  fig.height = 9,
  dpi = 300,
  ...
)
}
\arguments{
\item{data}{fold changes data}

\item{path}{saving path}

\item{cutoff}{significance cutoff}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}

\item{...}{ggplot2 extensions}
}
\description{
piePlot
}
