% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiamReduction.R
\name{plotDiamReduction}
\alias{plotDiamReduction}
\title{plotDiamReduction}
\usage{
plotDiamReduction(
  dataList,
  results,
  diam.method = c("pca", "opls", "tsne", "tsne_pca", "umap"),
  grouping.variables = NULL,
  dist.variables = NULL
)
}
\arguments{
\item{dataList}{raw data list of metabolome data}

\item{results}{results from diamntionality reduction}

\item{diam.method}{method of diamentionality reduction, pca, opls, tsne, tsne_pca, umap}

\item{grouping.variables}{list of grouping variables}

\item{dist.variables}{list of metabolites}
}
\value{
plotDiamReduction analyses in a list object.
   The object contains the following:\itemize{
    \item group.plot grouping variable plot
    \item distribution.plot distribution plot og listed metabolites.
  }
}
\description{
select and plot the distribution of selected method from compareDiamReduction function.
}
