% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetaboliteAlteration.R
\name{plotMetaboliteAlteration}
\alias{plotMetaboliteAlteration}
\title{plotMetaboliteAlteration}
\usage{
plotMetaboliteAlteration(
  data,
  path = NULL,
  cutoff = 0.01,
  lipid.class = TRUE,
  save = c("pdf", "svg", "png"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300
)
}
\arguments{
\item{data}{fold changes data}

\item{path}{saving path}

\item{cutoff}{significance cutoff}

\item{lipid.class}{wheather to plot lipid or not}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}
}
\value{
save object in defined path.
}
\description{
compute and plot distribution of signficantly altered metabolites per class.
}
