% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlots.R
\name{rocPlots}
\alias{rocPlots}
\title{rocPlots}
\usage{
rocPlots(
  dataList,
  group,
  path = NULL,
  var.imp = "PDAC",
  save = c("pdf", "svg", "png"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300
)
}
\arguments{
\item{dataList}{normalized data}

\item{group}{grouping variables}

\item{path}{saving path}

\item{var.imp}{disease for which ROC should be compared---should be one.}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}
}
\value{
summaryROC results table. rocPlots in save oblect in defined path.
}
\description{
compute and plot roc curves
}
