% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiroTest.R
\name{shapiroTest}
\alias{shapiroTest}
\title{shapiroTest}
\usage{
shapiroTest(dataList)
}
\arguments{
\item{dataList}{metabolome raw data expDataList}
}
\value{
results table
}
\description{
Shapiro-Wilk normality test
}
