% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banner.R
\name{banner}
\alias{banner}
\title{banner}
\usage{
banner(txt, symbol = "-")
}
\arguments{
\item{txt}{text to print in banner}

\item{symbol}{character separator (eg. "-", "=", "*", ....)}
}
\value{
tidy labels.
}
\description{
function to print tidy labels and separators
}
