% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlots.R
\name{boxPlots}
\alias{boxPlots}
\title{boxPlots}
\usage{
boxPlots(dataList, group = NULL, path = NULL)
}
\arguments{
\item{dataList}{raw metabolome data list from imputeTransformScale function.It need to have imputed.matrix and metadata.}

\item{group}{plotting grouping variable..should be 1}

\item{path}{saving path}
}
\value{
returns compiled pdf file.
}
\description{
function to plot box and violin plot from normalized metabolome data
}
