% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemicalMetadata.R
\docType{data}
\name{chemicalMetadata}
\alias{chemicalMetadata}
\title{chemical metadata for annotation of metabolon}
\format{
\describe{
        \item{CHEM_ID}{Chemical id}
        \item{MET_CHEM_NO}{Chemical id}
        \item{LIB_ID}{library id}
        \item{COMP_ID}{CCompound id}
        \item{CHRO_LIB_ENTRY_ID}{CHRO_LIB_ENTRY_ID}
        \item{SUPER_PATHWAY}{ONTOLOGY CLASS}
        \item{SUB_PATHWAY}{ONTOLOGY SUBCLASS}
        \item{PATHWAY_SORTORDER}{pathway sorting}
        \item{TYPE}{Named / unnamed}
        \item{INCHIKEY}{INCHIKEY}
        \item{SMILES}{SMILES}
        \item{CHEMICAL_NAME}{Chemical name}
        \item{PLOT_NAME}{Chemical name}
        \item{CAS}{CAS id}
        \item{CHEMSPIDER}{Chemspider id}
        \item{HMDB}{HMDB id}
        \item{KEGG}{KEGG id}
        \item{PUBCHEM}{PUBCHEM id}
        \item{PLATFORM}{used platform}
         }
}
\source{
{metabolon}
}
\usage{
chemicalMetadata
}
\description{
contains a dataframe with metadata
}
\examples{
data(chemicalMetadata)
}
\keyword{datasets}
