% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{correlationPlot}
\usage{
correlationPlot(
  dataList,
  h = 3,
  path = NULL,
  save = c("pdf", "svg", "png"),
  fig.width = 12,
  fig.height = 9
)
}
\arguments{
\item{dataList}{normalized data}

\item{h}{cluster size}

\item{path}{saving path}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}
}
\value{
correlation matrix as a list and pdf output in defined path folder.
}
\description{
function to plot intragroup correlation plot.
}
