% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{correlationPlot}
\usage{
correlationPlot(
  data = data,
  drop.var = NULL,
  h = 3,
  path = NULL,
  save = "pdf",
  fig.width = 12,
  fig.height = 9,
  ...
)
}
\arguments{
\item{data}{normalized data}

\item{drop.var}{grouping variables to drop}

\item{h}{cluster size}

\item{path}{saving path}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{...}{ggplot extensions}
}
\description{
correlationPlot
}
