% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionPlot.R
\name{distributionPlot}
\alias{distributionPlot}
\title{distributionPlot}
\usage{
distributionPlot(
  data,
  path = NULL,
  species = c("hsa", "mmu"),
  cutoff = 0.01,
  lipid.class = TRUE,
  data.type = c("MH", "Metabolon", "Others"),
  save = c("pdf", "svg", "png", "none"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300,
  Other_metadata = NULL
)
}
\arguments{
\item{data}{fold changes data}

\item{path}{saving path}

\item{species}{species to use "hsa" or "mmu"}

\item{cutoff}{cutoff of p value}

\item{lipid.class}{TRUE/FALSE to plot lipid classes}

\item{data.type}{select platform used, c("MH", "Metabolon", "Others")}

\item{save}{either "pdf", "svg" , "png", "none}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}

\item{Other_metadata}{dataframe with metadata....it must have  columns: Metabolite, Metabolite_Name, Ontology_Class, Ontology_Subclass}
}
\description{
plot distribution of all metabolites with fold changes and p-values.
}
