% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentScore.KEGG.R
\name{enrichmentScore.KEGG}
\alias{enrichmentScore.KEGG}
\title{enrichmentScore.KEGG}
\usage{
enrichmentScore.KEGG(
  species = c("hsa", "mmu"),
  ref.path = NULL,
  results,
  p.value.cutoff = 0.05,
  fold.changes.cutoff = 1.5,
  save = c("pdf", "svg", "png"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300
)
}
\arguments{
\item{species}{species to use "hsa" or "mmu"}

\item{ref.path}{saving path}

\item{results}{fold changes results}

\item{p.value.cutoff}{cutoff value of p.value}

\item{fold.changes.cutoff}{higher cutoff value of fold changes}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}
}
\value{
data.frame onject with enrichement.results and plots as save object in defined path.
}
\description{
enrichmentScore calualtion based on enrichement pathway analysis.
}
