% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leveneStat.R
\name{leveneStat}
\alias{leveneStat}
\title{leveneStat}
\usage{
leveneStat(
  group,
  dataList,
  location = c("mean", "median"),
  trim.alpha = 0.25,
  bootstrap = TRUE,
  num.bootstrap = 1000,
  kruskal.test = TRUE
)
}
\arguments{
\item{group}{grouping variable}

\item{dataList}{metabolome raw data expDataList}

\item{location}{mean or median, median is default}

\item{trim.alpha}{specify the percentage of trimmed mean}

\item{bootstrap}{whether to perform bootstrapping TRUE/FALSE}

\item{num.bootstrap}{number of bootstraping steps}

\item{kruskal.test}{whether to perform kruskal test TRUE/FALSE}
}
\value{
result table
}
\description{
computes Levene's stats for homogeneity of variance across groups
}
