% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeDat.R
\name{normalizeDat}
\alias{normalizeDat}
\title{normalizeDat}
\usage{
normalizeDat(
  data = data,
  confounders = NULL,
  stratifier = stratifier,
  fix.effect = NULL,
  random.effect = NULL
)
}
\arguments{
\item{data}{metabolome raw data}

\item{confounders}{list of confounders, NULL for mixed effect model}

\item{stratifier}{classifier variable of interest}

\item{fix.effect}{equation of fixed effect, NULL for fixed effect model}

\item{random.effect}{equation of random effects, NULL for fixed effect model}
}
\description{
Function to normalize data using fixed effect as well as mixed effects models.
}
