% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piePlot.R
\name{piePlot}
\alias{piePlot}
\title{piePlot}
\usage{
piePlot(
  data,
  path = NULL,
  cutoff = 0.05,
  lipid.class = TRUE,
  save = c("pdf", "svg", "png"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300
)
}
\arguments{
\item{data}{anova results obtained from normalizeDat.binary or normalizeDat function}

\item{path}{saving path}

\item{cutoff}{significance cutoff}

\item{lipid.class}{wheather to plot lipid or not}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}
}
\value{
output in save format in defined path
}
\description{
plot class distribution using pie charts
}
