% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piePlot.R
\name{piePlot}
\alias{piePlot}
\title{piePlot}
\usage{
piePlot(
  data,
  path = NULL,
  species = c("hsa", "mmu"),
  cutoff = 0.05,
  data.type = c("MH", "Metabolon", "Others"),
  lipid.class = TRUE,
  save = c("pdf", "svg", "png"),
  Other_metadata = NULL,
  fig.width = 12,
  fig.height = 9,
  dpi = 300
)
}
\arguments{
\item{data}{anova results obtained from normalizeDat.binary or normalizeDat function}

\item{path}{saving path}

\item{species}{species to use "hsa" or "mmu"}

\item{cutoff}{significance cutoff}

\item{data.type}{select platform used, c("MH", "Metabolon", "Others")}

\item{lipid.class}{wheather to plot lipid or not}

\item{save}{either "pdf", "svg" or "png"}

\item{Other_metadata}{dataframe with metadata....it must have  columns: Metabolite, Metabolite_Name, Ontology_Class, Ontology_Subclass}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}
}
\value{
output in save format in defined path
}
\description{
plot class distribution using pie charts
}
