% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetaboliteAlteration.R
\name{plotMetaboliteAlteration}
\alias{plotMetaboliteAlteration}
\title{plotMetaboliteAlteration}
\usage{
plotMetaboliteAlteration(
  data,
  path = NULL,
  species = c("hsa", "mmu"),
  cutoff = 0.01,
  lipid.class = TRUE,
  data.type = c("MH", "Metabolon", "Others"),
  save = c("pdf", "svg", "png"),
  fig.width = 12,
  fig.height = 9,
  dpi = 300,
  Other_metadata = NULL
)
}
\arguments{
\item{data}{fold changes data}

\item{path}{saving path}

\item{species}{species to use "hsa" or "mmu"}

\item{cutoff}{significance cutoff}

\item{lipid.class}{wheather to plot lipid or not}

\item{data.type}{select platform used, c("MH", "Metabolon", "Others")}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}

\item{Other_metadata}{dataframe with metadata....it must have  columns: Metabolite, Metabolite_Name, Ontology_Class, Ontology_Subclass}
}
\value{
save object in defined path.
}
\description{
compute and plot distribution of signficantly altered metabolites per class.
}
