% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSplitViolin.R
\name{plotSplitViolin}
\alias{plotSplitViolin}
\title{plotSplitViolin}
\usage{
plotSplitViolin(dataList, grouping.variable = NULL, markers = NULL)
}
\arguments{
\item{dataList}{raw data from metabolome study}

\item{grouping.variable}{grouping variable to plot..need to be binary or multinomial variable}

\item{markers}{list of markers to plots}
}
\value{
violin plot
}
\description{
plot split violin plot for comparison of two groups.
}
