% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlots.R
\name{rocPlots}
\alias{rocPlots}
\title{rocPlots}
\usage{
rocPlots(
  dataList,
  data.type = c("MH", "Metabolon", "Others"),
  species = c("hsa", "mmu"),
  group,
  path = NULL,
  var.imp = "PDAC"
)
}
\arguments{
\item{dataList}{normalized data}

\item{data.type}{select platform used, c("MH", "Metabolon", "Others")}

\item{species}{species to use "hsa" or "mmu"}

\item{group}{grouping variables}

\item{path}{saving path}

\item{var.imp}{disease for which ROC should be compared---should be one.}
}
\value{
summaryROC results table. rocPlots in save oblect in defined path.
}
\description{
compute and plot roc curves
}
