% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlots.R
\name{rocPlots}
\alias{rocPlots}
\title{rocPlots}
\usage{
rocPlots(
  data = data,
  group = group,
  drop.var = NULL,
  path = NULL,
  var.imp = "PDAC",
  save = "pdf",
  fig.width = 12,
  fig.height = 9,
  dpi = 300
)
}
\arguments{
\item{data}{normalized data}

\item{group}{grouping variables}

\item{drop.var}{drop un-necessary variables}

\item{path}{saving path}

\item{var.imp}{disease for which ROC should be compared}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}
}
\description{
rocPlots
}
