% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiroTest.R
\name{shapiroTest}
\alias{shapiroTest}
\title{shapiroTest}
\usage{
shapiroTest(data = data)
}
\arguments{
\item{data}{dataframe of which normality testing is necessary}
}
\description{
Shapiro-Wilk normality test
}
