% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{volcanoPlot}
\usage{
volcanoPlot(
  data = data,
  path = NULL,
  save = "pdf",
  fig.width = 12,
  fig.height = 9,
  dpi = 300,
  ...
)
}
\arguments{
\item{data}{fold changes data}

\item{path}{saving path}

\item{save}{either "pdf", "svg" or "png"}

\item{fig.width}{plot width not applicable for pdf}

\item{fig.height}{plot height not applicable for pdf}

\item{dpi}{dpi only applicable for png}

\item{...}{ggplot2 extensions}
}
\description{
Plot volcano plots
}
